requirements.txt
Django
python-dotenv
djangorestframework
djangorestframework-jwt
mysqlclient
